this is better

+++ custom/expressions/docker/src/configure-docker.sh
@@ -2,7 +2,15 @@
 
 TEMP_DIR=$(mktemp -d) &&
     cleanup() {
-	rm --recursive --force "${TEMP_DIR}" &&
+	if [ -z "$(ls -1 ${TEMP_DIR})" ]
+	then
+	    rm --recursive --force "${TEMP_DIR}" &&
+		true
+	else
+	    echo There were build errors. &&
+		echo "${TEMP_DIR}" &&
+		true
+	fi &&
 	    true
     } &&
     trap cleanup EXIT &&
@@ -14,10 +22,13 @@ TEMP_DIR=$(mktemp -d) &&
 	    WORK_DIR=$(mktemp -d "${TEMP_DIR}/XXXXXXXX") &&
 		cd "${WORK_DIR}" &&
 		cp --recursive "${STORE_DIR}/lib/${IMAGE}" "${WORK_DIR}" &&
-		nix-build "${WORK_DIR}/${IMAGE}/default.nix" &&
-		cat result | docker image load &&
+		date > date.log.txt 2>&1 &&
+		nix-build "${WORK_DIR}/${IMAGE}/default.nix" > nix-build.log.txt 2>&1 &&
+		cat result | docker image load > docker-image-load.log.txt 2>&1 &&
 		chown --recursive user:users "${WORK_DIR}" &&
 		chmod --recursive a+rwx "${WORK_DIR}" &&
+		cd "${TEMP_DIR}" &&
+		rm --recursive --force "${WORK_DIR}" &&
 		true
 	fi &&
 	    true
